BLAN-771 -- refactored the view of the events block
