+++ project/events/services.py
@@ -1,4 +1,3 @@
-import json
 import re
 from collections import OrderedDict
 from datetime import datetime
