+++ project/events/views/__init__.py
@@ -1,0 +1,31 @@
+from events.views.events_list import (
+    EventsList as EventsList,
+    EventsRelevantList as EventsRelevantList,
+    PopularEventsList as PopularEventsList,
+    UserEventsList as UserEventsList,
+    UserEventsRelevantList as UserEventsRelevantList,
+    UserParticipantEventsList as UserParticipantEventsList,
+    UserPlannedEventsList as UserPlannedEventsList,
+)
+
+from events.views.invites_to_event import (
+    BulkAcceptOrDeclineInvitesToEvent as BulkAcceptOrDeclineInvitesToEvent,
+    BulkAcceptOrDeclineRequestToParticipation as BulkAcceptOrDeclineRequestToParticipation,
+    InvitesToEventList as InvitesToEventList,
+    InviteUserToEvent as InviteUserToEvent,
+    RequestToParticipationsList as RequestToParticipationsList,
+)
+
+from events.views.manage_event import (
+    CreateEvent as CreateEvent,
+    UpdateEvent as UpdateEvent,
+    DeleteEvents as DeleteEvents,
+    GetEvent as GetEvent,
+)
+from events.views.leave_and_join_event import (
+    FanJoinToEvent as FanJoinToEvent,
+    FanLeaveFromEvent as FanLeaveFromEvent,
+    JoinToEvent as JoinToEvent,
+    LeaveFromEvent as LeaveFromEvent,
+    RemoveUserFromEvent as RemoveUserFromEvent,
+)
